Push comment options on the right
This commit pushes the comment options on the right. Refined Github
makes the gitHub interface better by removing useless features and by
organizing in an smarter way the other ones. Before this commit, if you
used an extension on the comment menu it was could be not properly
aligned. By removing the "margin-left" all options are pushed on the
right.

resolves #133