Replace babili with uglify-es
Lets us skip the extra babel-cli step.

Enables full source map output (otherwise not available because of
babel/babel#3940)

Which means we can have full source maps after the minification.

Which means that we get the real code even when debugging a published
version.

Which means that we can minify the JS fully and just submit the source
maps to Mozilla Addons (if they accept them)