Automatically focus "Confirm" in Mark all as read dialog
This is a simple enough change, as Chrome and Firefox both support the
"autofocus" attribute. It was just a matter of adding that attribute in
all the required cases.

I also added ".idea" to the .gitignore as that's what I use and I didn't
want to commit those files. If you'd prefer I revert that change I'd be
more than happy to do so.

Fixes #730