don't create tab on update
Extension updates happen at random times, unannounced. Opening a new tab steals focus and disrupts tasks. This is an awesome set-it-and-forget-it extension and it shouldn't be jumping in my face.

Simple fix: don't do it. A more interesting possibility would be to replace the new tab with a small overlay on the current page.