Remove minimum count in open-all-notifications.js
The "open all unread notifications in tabs" feature is very convenient,
as it requires a single click to be triggered.

When there's only one notification for a given repository,
the icon is not shown, which is understandable,
but that forces the same action to be done via a Ctrl+click,
or a right click to show the context menu and then "Open in a new tab".

Both options are less convenient than the button provided by this feature,
so this change simply removes the minimum of two notifications
to show the "open in new tabs" icon.

As a side effect, it also makes the notifications page more consistent,
since all notification blocks now have the same set of icons.