[bug] Fix null pointer error on GHE Projects page
For some reason a few pages in GHE, mainly projects, has different markup for the notification icon which doesn't have the `mail-class` list resulting in an error when we try to reference this `statusMark.classList.toggle('unread', hasUnread);` later.

GHE hasn't had notifications until a recent release, so likely a miss on their end.