Exclude directories from follow-file-renames
We cannot simply lookup the blob URL from the DOM because the `route`
field for trees would still be 'blob' and we would have to make a
`fetch` request and look for a redirect to determine whether the URL is
a 'tree' or a 'blob'.

There is another 'tree' URL in the DOM that points to 'Browse the
repository at the point in history' which is also unusable.