modify page navigation hotkeys to match GitHub
GitHub mapped pagination hotkeys `n` / `p` on pull request commit pages
overlapping the hotkeys set by this feature.

Since this project aims to keep as close to GitHub as possible and not
introduce new workflows for users, the feature is now modified to use
`n` and `p` as hotkeys for the feature ( hence changing its name ).

Also,
> Arrows feel wrong accessibility-wise.

https://github.com/sindresorhus/refined-github/pull/4691#issuecomment-907578663